顺序存储线性表分析及优化

+++ Object.cpp
@@ -6,22 +6,22 @@ using namespace std;
 namespace DTLib
 {
 
-void* Object::operator new(unsigned long long size) throw()
+void* Object::operator new (size_t size) throw()
 {
     return malloc(size);
 }
 
-void Object::operator delete(void *p)
+void Object::operator delete (void* p)
 {
     free(p);
 }
 
-void* Object::operator new[](unsigned long long size) throw()
+void* Object::operator new[] (size_t size) throw()
 {
     return malloc(size);
 }
 
-void Object::operator delete[](void *p)
+void Object::operator delete[] (void* p)
 {
     free(p);
 }
