+++ Object.h
@@ -7,9 +7,9 @@ namespace DTLib
 class Object
 {
 public:
-    void* operator new (unsigned long long size) throw();
+    void* operator new (size_t size) throw();
     void operator delete (void* p);
-    void* operator new[] (unsigned long long size) throw();
+    void* operator new[] (size_t size) throw();
     void operator delete[] (void* p);
     virtual ~Object() = 0;
 };
